requirements.txt
# ==========================================
# AWS y Bedrock
# ==========================================
boto3>=1.34.0
botocore>=1.34.0

# ==========================================
# Video Processing
# ==========================================
opencv-python>=4.9.0
numpy>=1.26.0

# ==========================================
# Utilities / CLI
# ==========================================
python-dotenv>=1.0.0
tqdm>=4.66.0
colorama>=0.4.6
